General cleanup
Cleaned up extention.ts
