+++ src/views/userConfig/index.js
@@ -47,7 +47,7 @@
         const r = parseInt(result[0]).toString(16).padStart(2, '0');
         const g = parseInt(result[1]).toString(16).padStart(2, '0');
         const b = parseInt(result[2]).toString(16).padStart(2, '0');
-        return '#' + (r + g + b).toUpperCase();
+        return '#' + (r + g + b + '4C').toUpperCase();
     }
 
     customColorInput.addEventListener('change', () => {
